Added a function to calculate gray maps in order to improve the loss against the imbalanced dataset
The dataset is imbalanced due to the amount of gray pixels in each image. As a result I wrote a function that uses a threshold to determine if the pixels are close to a shade of gray and I calculated the amount of gray and non-gray pixels present in the dataset.